introduce colored log levels using ansi_term
This commit adds a possibility to color the log levels using an ansi_term
crate for this purpose.

The change was intended for WriteLogger in mind, but as this feature will
be configurable and put in generic logging code, then other loggers can
benefit as well (if desired).

This way one can easily create a logfile with a similar colored look
as it is printed real-time on a terminal/console.

Tip: To view such a logfile with colors using 'less' you can use:
less -R /path/to/logfile